Complete `debug_build` workflow

+++ pyportable_installer/main_flow/step4/cleanup.py
@@ -12,13 +12,13 @@ def main(pyproj_file, additional_conf):
 
 
 def _cleanup_scaffold_files():
-    from ..step3.step3_3.create_launcher import thread
-    # wait for thread of 'generating exe from bat file' complete
-    if thread is not None:
-        thread.join()
-    # remove bat file
-    if exists(f := dst_model.launcher_bat):
-        os.remove(f)
+    from ..step3.step3_3.create_launcher import thread_of_bat_2_exe
+    # wait for thread_of_bat_2_exe of 'generating exe from bat file' complete
+    if thread_of_bat_2_exe is not None:
+        thread_of_bat_2_exe.join()
+        # remove bat file
+        if exists(f := dst_model.launcher_bat):
+            os.remove(f)
 
 
 def _generate_manifest(pyproj_file, file_o, additional_conf):
